+++ src/extract/extractor/tests.rs
@@ -41,6 +41,7 @@ mod exclusions;
 mod extraction_area;
 mod futures_returns;
 mod reexports;
+mod result_alias;
 mod serde;
 mod unsupported_generics;
 mod versioning;